assessor_land_exploration/California/Los Angeles County/pyproject.toml
[tool.poetry]
name = "assessor-land-exploration"
version = "0.1.0"
description = ""
authors = ["Julian Sotelo <juliansotelo@mail.fresnostate.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
openpyxl = "^3.1.2"
ipykernel = "^6.29.4"
plotly = "^5.21.0"
nbformat = "^5.10.4"
pyarrow = "^15.0.2"
fastparquet = "^2024.2.0"
pyspark = "^3.5.1"
polars = "^0.20.21"
xlsx2csv = "^0.8.2"
scipy = "^1.13.0"
dropbox = "^11.36.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
neattext = "^0.1.3"
cleantext = "^1.1.4"
censusgeocode = "^0.5.2"
requests-toolbelt = "0.10.1"
twine = "^5.0.0"
urllib3 = "1.26.15"
geopy = "^2.4.1"
pandas = "2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
